refactor: Improve useRooms hook functionality
